Added more to advanced concept
I did a bit more with it and made it so when the player dies from any enemy, the flying enemy will respawn back to its position. This is so the enemy cannot just fly through the levels to get back to the player from the other side of the level and make it weirdly hard to even move. I also added a little bit of a distance between the player and the enemy. When the enemy goes near the player, the enemy will stop and allow the player to continue at times depending on position. This is so the enemy does not engulf the player each time.